✨ Feature : 写真表示部分をコンポーネント化（後ほど修正する可能性大）
